mobile: unify navResetToHome, useExitToHome
