+++ apps/daimo-mobile/src/view/screen/AddPasskeyScreen.tsx
@@ -14,13 +14,14 @@ import { createPasskey } from "../../logic/passkey";
 import { useAccount } from "../../model/account";
 import { getAmountText } from "../shared/Amount";
 import { ButtonBig } from "../shared/Button";
+import { ScreenHeader } from "../shared/ScreenHeader";
+import Spacer from "../shared/Spacer";
 import {
-  ScreenHeader,
+  useDisableTabSwipe,
   useExitBack,
   useExitToHome,
-} from "../shared/ScreenHeader";
-import Spacer from "../shared/Spacer";
-import { useDisableTabSwipe, useNav } from "../shared/nav";
+  useNav,
+} from "../shared/nav";
 import { ss } from "../shared/style";
 import { TextCenter, TextError, TextLight, TextPara } from "../shared/text";
