+++ apps/daimo-mobile/src/view/screen/receive/DepositScreen.tsx
@@ -19,9 +19,10 @@ import { Account } from "../../../model/account";
 import { ButtonMed } from "../../shared/Button";
 import { CheckLabel } from "../../shared/Check";
 import { InfoBox } from "../../shared/InfoBox";
-import { ScreenHeader, useExitToHome } from "../../shared/ScreenHeader";
+import { ScreenHeader } from "../../shared/ScreenHeader";
 import { SegmentSlider } from "../../shared/SegmentSlider";
 import Spacer from "../../shared/Spacer";
+import { useExitToHome } from "../../shared/nav";
 import { color, ss, touchHighlightUnderlay } from "../../shared/style";
 import { TextBold, TextLight, TextPara } from "../../shared/text";
 import { useWithAccount } from "../../shared/withAccount";
